pull request fixes after code review

+++ MyGuide/Information/TicketsViewController.m
@@ -129,8 +129,8 @@ didSelectRowAtIndexPath: (NSIndexPath *)indexPath
     NSMutableArray *indexPaths   = [NSMutableArray new];
     NSArray *currentTicketsGroup = [self.ticketsGroups objectAtIndex:index];
     
-    long insertPos = index + 1;
-    for (int i = 0; i < [currentTicketsGroup count]; i++) {
+    NSInteger insertPos = index + 1;
+    for (NSInteger i = 0; i < [currentTicketsGroup count]; i++) {
         [indexPaths addObject:[NSIndexPath indexPathForRow:insertPos++ inSection:0]];
     }
     [self.tableView insertRowsAtIndexPaths: indexPaths
@@ -145,7 +145,7 @@ didSelectRowAtIndexPath: (NSIndexPath *)indexPath
 {
     NSMutableArray *indexPaths = [NSMutableArray new];
     
-    for (long i = index + 1; i <= index + [self childrenCount: index]; i++) {
+    for (NSInteger i = index + 1; i <= index + [self childrenCount: index]; i++) {
         [indexPaths addObject:[NSIndexPath indexPathForRow: i
                                                  inSection: 0]];
     }
